ci: Update CI/CD workflow to use 'master' branch
This commit updates the GitHub Actions workflow (`Async.Locks CI/CD`) to trigger on pushes to the `master` branch instead of the `main` branch.

The `branches` filter in the `on: push:` section of the workflow file is changed from `- main` to `- master`.

This modification ensures that the CI/CD pipeline is executed when changes are pushed to the `master` branch, aligning with the repository's branching strategy.