chore(ci): Suspend primary CI/CD pipeline temporarily
This commit comments out all lines in the `.github/workflows/ci-cd.yml` file. This action temporarily suspends the execution of the primary CI/CD pipeline.

This is likely done to allow for testing and validation of a new or modified CI/CD workflow defined in `test-ci-cd.yml` without interference from the existing pipeline. Once the testing of the new workflow is complete, the lines in this file will likely be uncommented to reactivate the primary CI/CD process.