feat: Update IAsyncLock to return IAsyncDisposable from AcquireAsync
This commit updates the `IAsyncLock` interface. The `AcquireAsync` method now consistently returns a `ValueTask<IAsyncDisposable>`. The disposable result is intended to be an `AsyncLockReleaser`, which is responsible for releasing the lock when disposed. The separate `ReleaseAsync` method is no longer part of the interface, as the release operation is now tied to the disposal of the releaser.